pyproject.toml
[tool.poetry]
name = "bank-prj"
version = "0.1.0"
description = "Hometask for course 2. Data for bank widget."
authors = ["Alexander Lukichev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
datetime = "^5.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.8.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
#ignore_missing_imports = true


[tool.black]
line-length = 119
exclude = '\.git'


[tool.isort]
line_length = 119
